# contracts/openapi/argus-api.yaml
openapi: 3.1.0
info:
  title: ARGUS Platform API
  version: 0.1.0
  description: GENERATED by scripts/contracts.py export from backend/app. Do not hand-edit.
paths:
...
  /api/v1/operators/{operator_id}/twin:
    get:
      tags:
      - twin
      summary: Get Operator Twin
      operationId: get_operator_twin_api_v1_operators__operator_id__twin_get
      security:
      - HTTPBearer: []
      parameters:
      - name: operator_id
        in: path
        required: true
        schema:
          type: string
          title: Operator Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalTwin'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sites/{site_id}/twin:
    get:
      tags:
      - twin
      summary: Get Site Twin
      operationId: get_site_twin_api_v1_sites__site_id__twin_get
      security:
      - HTTPBearer: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          title: Site Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteTwin'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/machines/{machine_id}/state:
    get:
      tags:
      - twin
      summary: Get Machine State
      operationId: get_machine_state_api_v1_machines__machine_id__state_get
      security:
      - HTTPBearer: []
      parameters:
      - name: machine_id
        in: path
        required: true
        schema:
          type: string
          title: Machine Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwinMachineState'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    GeoPoint:
      properties:
        lat:
          type: number
          maximum: 90.0
          minimum: -90.0
          title: Lat
        lon:
          type: number
          maximum: 180.0
          minimum: -180.0
          title: Lon
      type: object
      required:
      - lat
      - lon
      title: GeoPoint
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OperationalTwin:
      properties:
        operator_id:
          type: string
          title: Operator Id
        machine_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Machine Id
        task_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Task Id
        site_id:
          type: string
          title: Site Id
        as_of:
          type: string
          format: date-time
          title: As Of
        state:
          $ref: '#/components/schemas/TwinState'
        intelligence:
          $ref: '#/components/schemas/TwinIntelligence'
        freshness:
          $ref: '#/components/schemas/TwinFreshness'
      type: object
      required:
      - operator_id
      - site_id
      - as_of
      - state
      - intelligence
      - freshness
      title: OperationalTwin
      description: 'Human-Machine Operational Twin: state layer + intelligence layer (master §4).'
...
    SiteTwin:
      properties:
        site_id:
          type: string
          title: Site Id
        as_of:
          type: string
          format: date-time
          title: As Of
        operators:
          items:
            $ref: '#/components/schemas/OperationalTwin'
          type: array
          title: Operators
        machines:
          items:
            $ref: '#/components/schemas/TwinMachineState'
          type: array
          title: Machines
      type: object
      required:
      - site_id
      - as_of
      - operators
      - machines
      title: SiteTwin
...
    TaskTarget:
      properties:
        unit:
          type: string
          enum:
          - cycles
          - m3
          - tonnes
          - minutes
          title: Unit
        value:
          type: number
          exclusiveMinimum: 0.0
          title: Value
      type: object
      required:
      - unit
      - value
      title: TaskTarget
...
    TwinEnvironment:
      properties:
        weather:
          type: string
          enum:
          - CLEAR
          - CLOUDY
          - RAIN
          - STORM
          - FOG
          title: Weather
          default: CLEAR
        temperature_c:
          type: number
          title: Temperature C
          default: 30.0
        rain_mm_h:
          type: number
          minimum: 0.0
          title: Rain Mm H
          default: 0.0
        visibility:
          type: string
          enum:
          - GOOD
          - MODERATE
          - POOR
          title: Visibility
          default: GOOD
        soil:
          type: string
          enum:
          - DRY
          - WET
          - MUD
          - ROCKY
          title: Soil
          default: DRY
        zone_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone Id
        as_of:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: As Of
      type: object
      title: TwinEnvironment
      description: Conditions as recorded for a site or zone (name kept for API compatibility).
    TwinFreshness:
      properties:
        telemetry_age_s:
          anyOf:
          - type: number
          - type: 'null'
          title: Telemetry Age S
        environment_age_s:
          anyOf:
          - type: number
          - type: 'null'
          title: Environment Age S
      type: object
      title: TwinFreshness
    TwinIntelligence:
      properties:
        safety:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Safety
          description: From safety.event.raised
        risk:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Risk
          description: From prediction.risk.updated (evidence only)
        eta:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Eta
          description: From prediction.task_time.updated
        anomaly:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Anomaly
          description: From operator.anomaly.detected
        productivity:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Productivity
        environmental_difficulty:
          anyOf:
          - type: string
            enum:
            - LOW
            - MEDIUM
            - HIGH
          - type: 'null'
          title: Environmental Difficulty
      type: object
      title: TwinIntelligence
      description: 'Slots filled by other capabilities'' events. Nullable: the twin is valid without them.'
    TwinMachineState:
      properties:
        machine_id:
          type: string
          title: Machine Id
        machine_type:
          type: string
          title: Machine Type
        model:
          type: string
          title: Model
        telemetry_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Telemetry At
        operating_mode:
          anyOf:
          - type: string
            enum:
            - DIGGING
            - SWINGING
            - DUMPING
            - LOADING
            - HAULING
            - RETURNING
            - TRAVELLING
            - IDLE
            - 'OFF'
          - type: 'null'
          title: Operating Mode
        fuel_pct:
          anyOf:
          - type: number
          - type: 'null'
          title: Fuel Pct
        fuel_rate_lph:
          anyOf:
          - type: number
          - type: 'null'
          title: Fuel Rate Lph
        speed_kmh:
          anyOf:
          - type: number
          - type: 'null'
          title: Speed Kmh
        rpm:
          anyOf:
          - type: number
          - type: 'null'
          title: Rpm
        engine_hours:
          anyOf:
          - type: number
          - type: 'null'
          title: Engine Hours
        load_cycles:
          anyOf:
          - type: integer
          - type: 'null'
          title: Load Cycles
        idle_seconds:
          anyOf:
          - type: number
          - type: 'null'
          title: Idle Seconds
        temperature_c:
          anyOf:
          - type: number
          - type: 'null'
          title: Temperature C
        seatbelt:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Seatbelt
        location:
          anyOf:
          - $ref: '#/components/schemas/GeoPoint'
          - type: 'null'
        health:
          anyOf:
          - type: string
            enum:
            - OK
            - WARNING
            - CRITICAL
          - type: 'null'
          title: Health
      type: object
      required:
      - machine_id
      - machine_type
      - model
      title: TwinMachineState
    TwinOperatorState:
      properties:
        name:
          type: string
          title: Name
        status:
          type: string
          title: Status
        experience_level:
          type: string
          title: Experience Level
        certification_status:
          type: string
          title: Certification Status
        preferred_language:
          type: string
          title: Preferred Language
          default: en
        machine_familiarity:
          anyOf:
          - type: string
            enum:
            - LOW
            - MEDIUM
            - HIGH
          - type: 'null'
          title: Machine Familiarity
      type: object
      required:
      - name
      - status
      - experience_level
      - certification_status
      title: TwinOperatorState
    TwinSiteContext:
      properties:
        zone_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone Id
        zone_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone Type
        active_machines:
          type: integer
          title: Active Machines
          default: 0
        nearby_workers:
          anyOf:
          - type: integer
          - type: 'null'
          title: Nearby Workers
          description: Filled from the IoT hazard mesh; null until then
        congestion:
          anyOf:
          - type: string
            enum:
            - LOW
            - MEDIUM
            - HIGH
          - type: 'null'
          title: Congestion
          description: Filled from site operational intelligence; null until then
      type: object
      title: TwinSiteContext
    TwinState:
      properties:
        operator:
          $ref: '#/components/schemas/TwinOperatorState'
        machine:
          anyOf:
          - $ref: '#/components/schemas/TwinMachineState'
          - type: 'null'
        task:
          anyOf:
          - $ref: '#/components/schemas/TwinTaskState'
          - type: 'null'
        environment:
          anyOf:
          - $ref: '#/components/schemas/TwinEnvironment'
          - type: 'null'
        site:
          anyOf:
          - $ref: '#/components/schemas/TwinSiteContext'
          - type: 'null'
      type: object
      required:
      - operator
      title: TwinState
    TwinTaskState:
      properties:
        task_id:
          type: string
          title: Task Id
        title:
          type: string
          title: Title
        task_type:
          type: string
          enum:
          - EXCAVATION
          - TRENCHING
          - LOADING
          - HAULING
          - GRADING
          - BACKFILLING
          - OTHER
          title: Task Type
        status:
          type: string
          enum:
          - PLANNED
          - ASSIGNED
          - IN_PROGRESS
          - PAUSED
          - COMPLETED
          - CANCELLED
          title: Status
        zone_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone Id
        priority:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - CRITICAL
          title: Priority
        target:
          $ref: '#/components/schemas/TaskTarget'
        cycles_done:
          anyOf:
          - type: integer
          - type: 'null'
          title: Cycles Done
        progress_pct:
          anyOf:
          - type: number
          - type: 'null'
          title: Progress Pct
        planned_start:
          type: string
          format: date-time
          title: Planned Start
        planned_end:
          type: string
          format: date-time
          title: Planned End
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
        actual_start:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Actual Start
      type: object
      required:
      - task_id
      - title
      - task_type
      - status
      - priority
      - target
      - planned_start
      - planned_end
      title: TwinTaskState
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer